+++ frontend/src/routes/local/path.tsx
@@ -56,7 +56,7 @@ export default function SelectVideos() {
         </div>
         <div className="form-control">
           <label className="label">
-            <span className="label-text">Local path for your videos</span>
+            <span className="label-text">Location of your video</span>
           </label>
           <input
             required
