+++ frontend/src/routes/root.tsx
@@ -1,12 +1,158 @@
-import {LoaderFunction, json} from "react-router-dom"
+import {useStateMachine} from "little-state-machine"
+import React, {useEffect} from "react"
+import {
+  Link,
+  Outlet,
+  useLoaderData,
+  useNavigate,
+  useNavigation,
+} from "react-router-dom"
+import {HiCog, HiXMark} from "react-icons/hi2"
+import {resetForm} from "./actions"
+import {
+  FormStage,
+  LocalFilesFormStage,
+  LocalVideosFormState,
+  StashFormState,
+  StateHelpers,
+} from "../types/types"
+import Layout from "../components/Layout"
+import {getUrl} from "./stash/filter/root"
+import Steps from "../components/Steps"
 
-export const loader: LoaderFunction = async () => {
-  const response = await fetch("/api/stash/config")
-  if (response.ok) {
-    const config = await response.json()
-    return config
-  } else {
-    const error = await response.text()
-    throw json({error, request: "/api/stash/config"}, {status: 500})
-  }
+const StashSteps: React.FC<{state: StashFormState}> = ({state}) => {
+  return (
+    <Steps
+      currentStage={state.stage}
+      steps={[
+        {
+          stage: FormStage.SelectMode,
+          link: "/stash/mode",
+          content: "Choose mode",
+        },
+        {
+          stage: FormStage.SelectCriteria,
+          link: state.selectMode ? getUrl(state.selectMode) : "",
+          content: "Choose mode",
+        },
+        {
+          stage: FormStage.SelectMarkers,
+          link: "/stash/markers",
+          content: "Select markers",
+        },
+        {
+          stage: FormStage.Music,
+          link: "/stash/music",
+          content: "Music options",
+        },
+        {
+          stage: FormStage.VideoOptions,
+          link: "/stash/video-options",
+          content: "Video options",
+        },
+        {
+          stage: FormStage.PreviewClips,
+          link: "/stash/clips",
+          content: "Preview clips",
+        },
+        {
+          stage: FormStage.Wait,
+          link: "/stash/progress",
+          content: "Create video",
+        },
+      ]}
+    />
+  )
 }
+
+const LocalFileSteps: React.FC<{state: LocalVideosFormState}> = ({state}) => {
+  return (
+    <Steps
+      currentStage={state.stage}
+      steps={[
+        {
+          stage: LocalFilesFormStage.SelectPath,
+          link: "/local/path",
+          content: "Select file path",
+        },
+        {
+          stage: LocalFilesFormStage.ListVideos,
+          link: "/local/videos",
+          content: "Create markers",
+        },
+        {
+          stage: LocalFilesFormStage.Music,
+          link: "/stash/music",
+          content: "Music options",
+        },
+        {
+          stage: LocalFilesFormStage.VideoOptions,
+          link: "/stash/video-options",
+          content: "Video options",
+        },
+        {
+          stage: LocalFilesFormStage.PreviewClips,
+          link: "/stash/clips",
+          content: "Preview clips",
+        },
+        {
+          stage: LocalFilesFormStage.Wait,
+          link: "/stash/progress",
+          content: "Create video",
+        },
+      ]}
+    />
+  )
+}
+
+const RootLayout: React.FC = () => {
+  const {actions, state} = useStateMachine({resetForm})
+  const onReset = () => {
+    if (
+      confirm(
+        "Are you sure you want to reset the form and return to the start?"
+      )
+    ) {
+      actions.resetForm()
+      navigate("/")
+    }
+  }
+  const navigate = useNavigate()
+  const navigation = useNavigation()
+  const isLoading = navigation.state === "loading"
+  const config = useLoaderData()
+  const configExists = config !== null
+  const isStashMode = StateHelpers.isStash(state.data)
+
+  useEffect(() => {
+    if (!configExists && isStashMode) {
+      navigate("/stash/config")
+    }
+  }, [configExists, isStashMode])
+
+  return (
+    <Layout isLoading={isLoading}>
+      <section className="py-4 flex flex-col">
+        <h1 className="text-5xl font-bold mb-4 text-center">ClipMash</h1>
+        <div className="self-center flex gap-2 mb-4">
+          <button onClick={onReset} className="btn btn-sm btn-error">
+            <HiXMark className="w-5 h-5 mr-2" />
+            Reset
+          </button>
+
+          <Link to="/config" className="btn btn-sm btn-info">
+            <HiCog  className="w-5 h-5 mr-2 "/>
+            Settings
+          </Link>
+        </div>
+        {StateHelpers.isStash(state.data) && <StashSteps state={state.data} />}
+        {StateHelpers.isLocalFiles(state.data) && (
+          <LocalFileSteps state={state.data} />
+        )}
+        <Outlet />
+      </section>
+    </Layout>
+  )
+}
+
+export default RootLayout
